refactor: Extract animation state into a dedicated component
That will make possible to move the animation definition in the assets.

And, it also makes easier (for benimator and the user) to react on
change of the animation component, since it will no longer be updated just because
the state has changed.

BREAKING CHANGE: The system label enum is renamed to
AnimationPostUpdateSystem, as the animation now runs during the
post-update phase